refactor code to use variables from cdk.json. Refactor eks_cluster.py into more functions.
