+++ cluster-bootstrap/cdk.json
@@ -1,4 +1,49 @@
   
 { 
-    "app": "python eks_cluster.py"
+    "app": "python3 app.py",
+    "context": {
+        "account": "",
+        "region": "",
+        "eks_version": "",
+        "node_type": "m5.large",
+        "node_quantity": "3",
+        "node_disk_size": "20",
+        "node_ami_version": "1.19.6-20210504",
+        "deploy_bastion": "True",
+        "basiton_node_type": "t3.large",
+        "basiton_disk_size": "20",
+        "deploy_client_vpn": "False",
+        "client_certificate_arn": "",
+        "server_certificate_arn": "",
+        "vpn_client_cidr_block": "",
+        "create_new_vpc": "",
+        "vpc_cidr": "10.0.0.0/16",
+        "vpc_cidr_mask_public": "24",
+        "vpc_cidr_mask_private": "24",
+        "existing_vpc_name": "",
+        "create_new_cluster_admin_role": "",
+        "existing_role_arn": "",
+        "deploy_aws_lb_controller": "",
+        "deploy_external_dns": "",
+        "deploy_managed_elasticsearch": "",
+        "es_data_nodes": "1",
+        "es_data_node_instance_type": "r5.large.elasticsearch",
+        "es_master_nodes": "0",
+        "es_master_node_instance_type": "r5.large.elasticsearch",
+        "es_enabled": "True",
+        "es_volume_type": "GP2",
+        "es_volume_size": "10",
+        "deploy_kube_prometheus_operator": "",
+        "deploy_aws_ebs_csi": "",
+        "deploy_aws_efs_csi": "",
+        "deploy_opa_gatekeeper": "",
+        "deploy_gatekeeper_policies": "",
+        "gatekeeper_policies_git_url": "ssh://git@github.com/aws-quickstart/quickstart-eks-cdk-python",
+        "gatekeeper_policies_git_branch": "main",
+        "gatekeeper_policies_git_path": "gatekeeper-policies",
+        "deploy_cluster_autoscaler": "",
+        "deploy_metrics_server": "",
+        "deploy_calico_np": "",
+        "deploy_ssm_agent": ""
+    }
 }
